+++ common/routes/Home/components/VideoBlock.js
@@ -25,7 +25,8 @@ class VideoBlock extends PureComponent {
 
   onSelectThumb (index) {
     this.setState({selectedIndex: index});
-    window.scroll(0, 3150);
+    // 2330px = MainVideo 到底下的高度
+    window.scroll(0, document.body.scrollHeight - 2330);
   }
 
   prev () {