+++ server/index.prod.js
@@ -26,6 +26,7 @@ import sitemap from './sitemap';
 const defaultServerConfig = configLib.get('server');
 const webApiServer = configLib.get('webApiServer');
 const memberApiServer = configLib.get('memberApiServer');
+const imgApiServer = configLib.get('imgApiServer');
 const headers = configLib.get('headers');
 const region = configLib.get('region');
 
@@ -73,7 +74,8 @@ export const createServer = (config) => {
           query: req.query,
           user
         },
-        memberServ: memberApiServer
+        memberServ: memberApiServer,
+        imgServ: imgApiServer
       }
     });
     const routes = createRoutes(store);