+++ CHANGELOG.md
@@ -5,13 +5,14 @@
 而項目則基於 [Semantic Versioning](http://semver.org/lang/zh-TW/).
 
 ## [Unreleased]
-### Changed
-- 隱藏作者圖片，避免爬蟲因為抓不到主圖而抓到作者圖 @esbb48
 
 ## 1.0.27 - 2017-08-29
 ### Added
 - 增加廣告版位 `/5799246/column_300x250_amm_3` @esbb48
 - 加上 `ad=0` @esbb48
+
+### Changed
+- 隱藏作者圖片，避免爬蟲因為抓不到主圖而抓到作者圖 @esbb48
 
 ## 1.0.26 - 2017-08-25
 ### Fixed